_setup_env.sh: Allow HOSTCC to be overridden even with `LLVM=1`
Some kind of compilers such as Snapdragon LLVM, doesn't support compiling host
binaries. This change is required to allow hosts binaries to be compiled with
a different compiler than what will be used by default.

Note that on master-kernel-build-2022 this is applied to _setup_env.sh instead
of build.sh due to transition to Kleaf.

This works even on hermetic setup.

Signed-off-by: Albert I <kras@raphielgang.org>