build_utils.sh: remove unused modules for the first stage more precisely
After collecting first stage modules, we also need to remove unused modules
in kernel path of first stage. This change can more precisely select
unused modules and remove them. For example, when we have b-a.ko and a.ko,
the current code will remove b-a.ko and a.ko at the same time although
the first stage needs b-a.ko not a.ko. Default code is to use phrase
to match module's name, so the b-a.ko is same with a.ko. This change will
insert escaped directory boundary (.\/.) in front of every line in
unused modules list to match the whole kernel path's modules in
the first stage. This can match entire module's name so that b-a.ko
is not same with a.ko.

Change-Id: Id484caea2ba1b825cce59458dc69ef1f6fb6098d
Signed-off-by: Song Xue <quic_songxue@quicinc.com>